requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
tqdm>=4.65.0
PyYAML>=6.0

# Model architectures
timm>=0.9.0
transformers>=4.30.0

# Adversarial attacks
torchattacks>=3.4.0

# Diffusion models
diffusers>=0.20.0
accelerate>=0.20.0

# Metrics
lpips>=0.1.4
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: HuggingFace authentication
huggingface_hub>=0.16.0
